Move Menu Class to the service package
